requirements.txt
amqp==1.4.6
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.20
boto==2.38.0
celery==3.1.18
decorator==4.0.2
dj-database-url==0.3.0
Django==1.8.4
django-admin-sortable==2.0.1
django-filter==0.11.0
django-mptt==0.7.4
django-ses==0.7.0
djangorestframework==3.2.4
easy-thumbnails==2.2
feedparser==5.2.1
FeinCMS==1.11.1
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
kombu==3.0.26
Markdown==2.6.2
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
Pillow==2.9.0
psycopg2==2.6.1
pytz==2015.6
reportlab==3.2.0
requests==2.7.0
simplegeneric==0.8.1
traitlets==4.0.0
WebOb==1.4.1
wheel==0.24.0
XlsxWriter==0.7.3
